feat: add TaskListOutputDTO
