+++ src/main/java/task/dto/TaskListOutputDTO.java
@@ -1,0 +1,14 @@
+package task.dto;
+
+import java.util.List;
+
+public record TaskListOutputDTO(
+        List<OutputTaskDTO> tasks,
+        String outputState
+) implements OutputDTO {
+
+    @Override
+    public String getOutputState() {
+        return outputState;
+    }
+}
