+++ src/test/java/task/service/TaskServiceTest.java
@@ -1,7 +1,7 @@
 package task.service;
 
 import common.exception.DataAccessException;
-import common.exception.DataAccessException;
+import common.exception.InvalidTaskIDException;
 import common.exception.TaskNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,15 +9,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import task.dto.*;
 import task.enums.Priority;
 import task.enums.TaskState;
+import task.mapper.TaskToDTOMapper;
 import task.model.Task;
 import task.model.TaskBuilder;
 import task.repository.TaskRepository;
 
-import java.time.LocalDateTime;
 import java.util.Collections;
-import java.util.List;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,7 +26,10 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-public class TaskServiceTest {
+public class TaskServiceGetTaskByIDTest {
+
+    @Mock
+    private TaskToDTOMapper mapper;
 
     @Mock
     private TaskRepository repository;
@@ -33,126 +37,308 @@ public class TaskServiceTest {
     @InjectMocks
     private TaskService service;
 
+    /* =============================== create test methods =====================================*/
+    @Test
+    @DisplayName("Should return success message when task is created successfully")
+    void createTask_validTask_returnsSuccessMessage() {
+        TaskDTO inputDTO = new TaskDTO("New Task", "Task Description", "", "medium");
+
+        Task taskToCreate = TaskBuilder.newTask()
+                .withTitle("New Task")
+                .withDescription("Task Description")
+                .build();
+
+        Task createdTask = TaskBuilder.newTask()
+                .withTitle("New Task")
+                .withDescription("Task Description")
+                .build();
+        createdTask.setId("123456");
+
+        OutputTaskDTO expectedOutput = new OutputTaskDTO(
+                "123456",
+                "New Task",
+                "Task Description",
+                null,
+                "20-02-2026 10:30:00",
+                "MEDIUM",
+                "NOT_COMPLETED",
+                "New Task created"
+        );
+
+        when(mapper.dtoToTask(inputDTO)).thenReturn(taskToCreate);
+        when(repository.create(any(Task.class))).thenReturn(createdTask);
+        when(mapper.taskToDto(createdTask, "New Task created")).thenReturn(expectedOutput);
+
+        OutputDTO result = service.createTask(inputDTO);
+
+        assertEquals("New Task created", result.getOutputState());
+        assertInstanceOf(OutputTaskDTO.class, result);
+
+        verify(mapper).dtoToTask(inputDTO);
+        verify(repository).create(any(Task.class));
+        verify(mapper).taskToDto(createdTask, "New Task created");
+    }
+
+    @Test
+    @DisplayName("Should throw IllegalArgumentException when TaskDTO is null")
+    void createTask_nullTask_throwsIllegalArgumentException() {
+        TaskDTO nullTaskDTO = null;
+
+        IllegalArgumentException exception = assertThrows(
+                IllegalArgumentException.class,
+                () -> service.createTask(nullTaskDTO)
+        );
+
+        assertEquals("CreateTask: Task cannot be null", exception.getMessage());
+        verifyNoInteractions(repository);
+        verifyNoInteractions(mapper);
+    }
+
+    @Test
+    @DisplayName("Should return error message when repository throws IllegalArgumentException")
+    void createTask_invalidDataFormat_returnsErrorMessage() {
+        TaskDTO inputDTO = new TaskDTO("Task", "new description", "", "medium");
+
+        Task taskToCreate = TaskBuilder.newTask()
+                .withTitle("Task")
+                .withDescription("new description")
+                .build();
+
+        when(mapper.dtoToTask(inputDTO)).thenReturn(taskToCreate);
+        doThrow(new IllegalArgumentException("The document can't be empty"))
+                .when(repository).create(any(Task.class));
+
+        OutputDTO result = service.createTask(inputDTO);
+
+        assertInstanceOf(ErrorOutputDTO.class, result);
+        assertEquals("Something go wrong with data format", result.getOutputState());
+
+        verify(mapper).dtoToTask(inputDTO);
+        verify(repository).create(any(Task.class));
+        verify(mapper, never()).taskToDto(any(), any());
+    }
+
+    @Test
+    @DisplayName("Should return error message when database infrastructure failure occurs")
+    void createTask_dataAccessException_returnsErrorMessage() {
+        TaskDTO inputDTO = new TaskDTO("New Task", "new description", "", "high");
+
+        Task taskToCreate = TaskBuilder.newTask()
+                .withTitle("New Task")
+                .withDescription("new description")
+                .build();
+
+        when(mapper.dtoToTask(inputDTO)).thenReturn(taskToCreate);
+        doThrow(new DataAccessException("MongoDB connection failed"))
+                .when(repository).create(any(Task.class));
+
+        OutputDTO result = service.createTask(inputDTO);
+
+        assertInstanceOf(ErrorOutputDTO.class, result);
+        assertEquals("Error raised during task creation. Try again.", result.getOutputState());
+
+        verify(mapper).dtoToTask(inputDTO);
+        verify(repository).create(any(Task.class));
+        verify(mapper, never()).taskToDto(any(), any());
+    }
+    /* =============================== getTaskByID test methods =====================================*/
     @Test
     @DisplayName("GetTaskById must return the task with the requested ID")
     void testGetTaskById_Positive() {
-
         String id = "69949f595f811f0d2276b457";
 
         Task mockTask = TaskBuilder.newTask()
                 .withTitle("Testing Title")
-                .withDescription("Testing Description")
+                .withDescription("Description")
                 .build();
-
         mockTask.setId(id);
 
-        when(repository.getById(id)).thenReturn(Optional.of(mockTask));
+        OutputTaskDTO expectedOutput = new OutputTaskDTO(
+                id,
+                "Testing Title",
+                "Description",
+                null,
+                "20-02-2026 10:30:00",
+                "MEDIUM",
+                "NOT_COMPLETED",
+                "Task retrieved successfully"
+        );
 
-        Task result = service.getTaskById(id);
+        TaskIdDTO idDTO = new TaskIdDTO(id);
+
+        when(repository.getById(id)).thenReturn(Optional.of(mockTask));
+        when(mapper.taskToDto(mockTask, "Task retrieved successfully")).thenReturn(expectedOutput);
+
+        OutputDTO result = service.getTaskById(idDTO);
 
         assertNotNull(result, "Expected a task but received null");
-        assertEquals(id, result.getId(), "The Service should return the task with the requested ID");
+        assertInstanceOf(OutputTaskDTO.class, result);
+        assertEquals(id, ((OutputTaskDTO) result).id(), "The Service should return the task with the requested ID");
+        assertEquals("Task retrieved successfully", result.getOutputState());
 
-        verify(repository, times(1)).getById(id); // esto verifica que se llamo al repository
+        verify(repository, times(1)).getById(id);
+        verify(mapper, times(1)).taskToDto(mockTask, "Task retrieved successfully");
     }
 
     @Test
-    @DisplayName("It must throw an exception if the task does not exist in the database.")
+    @DisplayName("Must return ErrorOutputDTO if the task does not exist in the database")
     void testGetTaskById_Negative() {
         String notExistentId = "empty/null/wrong";
+        TaskIdDTO idDTO = new TaskIdDTO(notExistentId);
 
         when(repository.getById(notExistentId)).thenReturn(Optional.empty());
 
-        assertThrows(TaskNotFoundException.class, () -> {
-            service.getTaskById(notExistentId);
-        });
+        OutputDTO result = service.getTaskById(idDTO);
+
+        assertNotNull(result);
+        assertInstanceOf(ErrorOutputDTO.class, result);
+        assertEquals("Task not found", result.getOutputState());
+
+        verify(repository, times(1)).getById(notExistentId);
+        verify(mapper, never()).taskToDto(any(), any());
+    }
+    /* =============================== remove test methods =====================================*/
+
+    @Test
+    @DisplayName("Should return a success message when the task is deleted correctly")
+    void removeTask_taskExists() {
+        String validId = "69949f595f811f0d2276b457";
+
+        doNothing().when(repository).remove(validId);
+        OutputDTO result = service.removeTask(new TaskIdDTO(validId));
+
+        assertEquals("Task (" + validId + ") successfully deleted.", result.getOutputState());
+        verify(repository).remove(validId);
     }
 
     @Test
-    @DisplayName("It should call repository.modify when the task is valid")
-    void updateTask_ValidTaskWithId_ReturnsSuccessMessage() {
-        Task mockTask = TaskBuilder.newTask()
-                .withTitle("Testing Title")
-                .withDescription("Testing Description")
-                .build();
-        mockTask.setId("69949f595f811f0d2276b457");
+    @DisplayName("Should return a not found message when the task does not exist")
+    void removeTask_taskNotFound() {
+        String validId = "69949f595f811f0d2276b457";
+
+        doThrow(new TaskNotFoundException(validId)).when(repository).remove(validId);
+        OutputDTO result = service.removeTask(new TaskIdDTO(validId));
+
+        assertEquals("Task (" + validId + ") does not exist.", result.getOutputState());
+        verify(repository).remove(validId);
+    }
+
+    @Test
+    @DisplayName("Should return an invalid format message when the ID is malformed")
+    void removeTask_invalidIdFormat() {
+        String invalidId = "this-is-not-an-object-id";
+
+        doThrow(new InvalidTaskIDException("Invalid format: " + invalidId))
+                .when(repository).remove(invalidId);
+        OutputDTO result = service.removeTask(new TaskIdDTO(invalidId));
+
+        assertEquals("Invalid id format.", result.getOutputState());
+        verify(repository).remove(invalidId);
+    }
+
+    @Test
+    @DisplayName("Should return an error message when a database infrastructure failure occurs")
+    void removeTask_dataAccessException() {
+        String validId = "69949f595f811f0d2276b457";
+
+        doThrow(new DataAccessException("MongoDB failure")).when(repository).remove(validId);
+        OutputDTO result = service.removeTask(new TaskIdDTO(validId));
+
+        assertEquals("Error raised during task deletion. Try again.", result.getOutputState());
+        verify(repository).remove(validId);
+    }
+
+    /* =============================== update test methods =====================================*/
+
+    @Test
+    @DisplayName("updateTask_ValidRequest_ReturnsOutputTaskDTO")
+    void updateTask_ValidRequest_ReturnsOutputTaskDTO() {
+        String id = "69949f595f811f0d2276b457";
+        TaskUpdateDTO inputDTO = new TaskUpdateDTO(id, "Title", "Desc", "2026-12-31", "HIGH");
+
+        // 1. Usamos la clase exacta que devuelve tu mapper
+        OutputTaskDTO expectedResponse = new OutputTaskDTO(
+                id, "Title", "Desc", "2026-12-31", "2026-02-21", "HIGH", "TODO",
+                "Task with _id " + id + " updated successfully"
+        );
+
+        // 2. Configuramos el mock para que devuelva el objeto correcto
+        when(repository.getById(id)).thenReturn(Optional.of(mock(Task.class)));
+        when(mapper.dtoUpdateToTask(any(), any())).thenReturn(mock(Task.class));
+        when(repository.modify(any())).thenReturn(mock(Task.class));
+
+        // IMPORTANTE: El retorno del mapper debe coincidir con expectedResponse
+        when(mapper.taskToDto(any(Task.class), anyString())).thenReturn(expectedResponse);
+
         // When
-        String result = service.updateTask(mockTask);
+        OutputDTO result = service.updateTask(inputDTO);
 
         // Then
-        assertEquals("Task with _id 69949f595f811f0d2276b457 updated successfully", result);
-        verify(repository, times(1)).modify(mockTask);
-
+        assertNotNull(result);
+        // Si OutputTaskDTO tiene los datos correctos, el test pasará
+        assertEquals(expectedResponse, result);
     }
 
     @Test
-    @DisplayName("It should throw IllegalArgumentException when the input task is null")
-    void updateTask_NullTask_ReturnsErrorMessage() {
-
-        String result = service.updateTask(null);
-
-        assertEquals("Error: Task cannot be null", result);
-    }
-
-    @Test
-    @DisplayName("It should throw TaskNotFoundException and preserve cause when repository fails")
-    void updateTask_IdNotFoundInRepository_ReturnsPersistenceErrorMessage() {
-        // Given
-        Task mockTask = TaskBuilder.newTask()
-                .withTitle("Ghost Task")
-                .withDescription("Non-existent task description")
-                .build();
-        mockTask.setId("999999999999999999999999");
-
-        // Simulamos que el DAO/Repository lanza la excepción técnica
-        doThrow(new DataAccessException("ID not found"))
-                .when(repository).modify(any(Task.class));
-
+    @DisplayName("Should throw IllegalArgumentException when the input DTO is null")
+    void updateTask_NullInput_ThrowsIllegalArgumentException() {
         // When & Then
-        String result = service.updateTask(mockTask);
-
-        // Verificamos congruencia de mensajes y encadenamiento (Stack Trace)
-        assertTrue(result.contains("Persistence error: ID not found"));
+        assertThrows(IllegalArgumentException.class, () -> service.updateTask(null));
     }
 
     @Test
-    @DisplayName("It should return error message when the task has no ID")
-    void updateTask_MissingId_ReturnsErrorMessage() {
-        // Given: Una tarea construida pero sin setearle el ID
-        Task taskWithoutId = TaskBuilder.newTask()
-                .withTitle("No ID Task")
-                .withDescription("Non-existent task description")
-                .build();
-
-        // When
-        String result = service.updateTask(taskWithoutId);
-
-        // Then
-        assertEquals("Error: Cannot update a task without an ID", result);
-        // Verificamos que NI SIQUIERA se intentó llamar al repositorio
-        verify(repository, never()).modify(any());
-    }
-
-    @Test
-    @DisplayName("It should return unexpected error message for non-data exceptions")
-    void updateTask_UnexpectedException_ReturnsGeneralErrorMessage() {
+    @DisplayName("updateTask_IdNotFound_ReturnsErrorOutputDTO")
+    void updateTask_IdNotFound_ReturnsErrorOutputDTO() {
         // Given
-        Task mockTask = TaskBuilder.newTask()
-                .withTitle("Testing Title")
-                .withDescription("Testing Description")
-                .build();
-        mockTask.setId("69949f595f811f0d2276b457");
+        String id = "non-existent";
+        TaskUpdateDTO inputDTO = new TaskUpdateDTO(id, "Title", "Desc", null, null);
 
-        // Simulamos un error que NO sea de persistencia (ej. un RuntimeException cualquiera)
-        doThrow(new RuntimeException("Fatal system crash"))
-                .when(repository).modify(any(Task.class));
+        when(repository.getById(id)).thenReturn(Optional.empty());
 
         // When
-        String result = service.updateTask(mockTask);
+        OutputDTO result = service.updateTask(inputDTO);
 
         // Then
-        assertTrue(result.contains("Unexpected error: Fatal system crash"));
+        assertTrue(result instanceof ErrorOutputDTO);
+        // Usamos getOutputState() que es el método que implementa tu Record
+        assertTrue(result.getOutputState().contains("not found"));
     }
+
+    @Test
+    @DisplayName("updateTask_RepositoryFails_ReturnsErrorOutputDTO")
+    void updateTask_RepositoryFails_ReturnsErrorOutputDTO() {
+        // Given
+        String id = "69949f595f811f0d2276b457";
+        TaskUpdateDTO inputDTO = new TaskUpdateDTO(id, "Title", "Desc", null, null);
+        Task existingTask = mock(Task.class);
+
+        when(repository.getById(id)).thenReturn(Optional.of(existingTask));
+        when(mapper.dtoUpdateToTask(any(), any())).thenReturn(mock(Task.class));
+        when(repository.modify(any())).thenThrow(new DataAccessException("DB Connection Lost"));
+
+        // When
+        OutputDTO result = service.updateTask(inputDTO);
+
+        // Then
+        assertTrue(result instanceof ErrorOutputDTO);
+        assertTrue(result.getOutputState().contains("Persistence error"));
+    }
+
+    @Test
+    @DisplayName("updateTask_GeneralException_ReturnsErrorOutputDTO")
+    void updateTask_GeneralException_ReturnsErrorOutputDTO() {
+        // Given
+        TaskUpdateDTO inputDTO = new TaskUpdateDTO("123", "Title", "Desc", null, null);
+        when(repository.getById(any())).thenThrow(new RuntimeException("Fatal Crash"));
+
+        // When
+        OutputDTO result = service.updateTask(inputDTO);
+
+        // Then
+        assertTrue(result instanceof ErrorOutputDTO);
+        assertTrue(result.getOutputState().contains("Unexpected error"));
+    }
+    /* =============================== listTasksByStatus test methods =====================================*/
 
     @Test
     @DisplayName("It should return a formatted list when completed tasks exist")
